% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RunPRMD.R
\name{patlist.display}
\alias{patlist.display}
\title{Display patient records}
\usage{
patlist.display(patlist, n.dose, n.cycle)
}
\arguments{
\item{patlist}{A list of the patient treatment records, which must contains
the following variables: \describe{ \item{PatID}{denotes the patient ID
where the elements are specified by cohort and subject number. For
example, "cohort2subject3" denotes the third subject in the second
cohort} \item{dose}{records the dose level assigned for each patient
through the whole treatment} \item{cycle}{shows the treatment cycle
information of each record} \item{nTTP}{records the corresponding nTTP
score.} \item{dlt}{indicates whether a DLT event is observed or not?}}}

\item{n.dose}{The number of dose in the study}

\item{n.cycle}{The number of cycle in the study}
}
\description{
Display patient records in a human-readable table. Each cell contains 3
values including observed nTTP, DLT, and dose assignment. The higher the
dose, the warmer the cell background color is. The black color of the
records indicates DLT equals 1.
}
