% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patlist_sim}
\alias{patlist_sim}
\title{A list of patient information}
\format{An object of class \code{list} of length 6.}
\usage{
patlist_sim
}
\value{
\item{PatID}{denotes the patient ID where the elements are specified by cohort
and subject number. For example, "cohort2subject3" denotes the third subject
in the second cohort}

\item{dose}{records the dose assigned for each patient on different cycles}

\item{cycle}{shows the treatment cycle}

\item{nTTP}{records th corresponding nTTP score.}

\item{dlt}{indicates whether a DLT event is observed or not?}

\item{efficacy}{provides the continuous efficacy for each cycle. The range of
efficacy measure is (0, 1).}
}
\description{
The data "patlist_sim" is a trial generated by "SimPRMD". The model utilized
for generating trial is a 3-stage model with individualized dose modification.
The MTD = 4 and the MED = 1 since the dose-efficacy pattern when generating
the trial was set to be flat. The dose-toxicity trend and efficacy-cycle trend
are flat. There are in total 12 cohorts with 3 patients in each cohort. This
also serves as an example of the input data of function \code{\link{RunPRMD}}.
}
\keyword{datasets}
