% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RunPRMD.R
\name{plot.RunPRMD}
\alias{plot.RunPRMD}
\title{Plot nTTP and efficacy boxplots of a RunPRMD object}
\usage{
\method{plot}{RunPRMD}(x, ..., select_cycle = x$cycles)
}
\arguments{
\item{x}{RunPRMD object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}

\item{select_cycle}{A vector indication the cycle in the boxplot. Default
is \code{cycle} of \code{x}.}
}
\description{
Plot nTTP boxplots of a RunPRMD object. Plot efficacy boxplots when
implementing \code{RunPRMD} with option \code{effcy.flag == TRUE}.
}
\examples{
## Check ?RunPRMD for example

}
