% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SimPRMD.R
\name{plot.SimPRMD}
\alias{plot.SimPRMD}
\title{Plots of a SimPRMD object}
\usage{
\method{plot}{SimPRMD}(x, ..., title.add = TRUE)
}
\arguments{
\item{x}{SimPRMD object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}

\item{title.add}{controls whether there is a title on plots or not.}
}
\description{
Plot the predictive probability of nTTP < target toxicity for all cycles
and doses , the mean nTTP vs cycle1 and cycle > 2 for all doses of a
SimPRMD object. Plot median treatment duration boxplot along with the DLT
drop off rate when implementing \code{\link{SimPRMD}} with option
\code{DLT.drop.flag = TRUE}.
}
\examples{
## Check ?SimPRMD for example

}
