% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.RunPRMD.R
\name{print.summary.RunPRMD}
\alias{print.summary.RunPRMD}
\title{Displays a useful description of a summary.RunPRMD object}
\usage{
\method{print}{summary.RunPRMD}(x, ...)
}
\arguments{
\item{x}{summary.RunPRMD object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays a useful description of a summary.RunPRMD object. Call by
\code{link{summary.RunPRMD}}. Check \code{link{summary.RunPRMD}} for the
details of the print information.
}
