% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SimPRMD.R
\name{print.summary.SimPRMD}
\alias{print.summary.SimPRMD}
\title{Displays a useful description of a summary.SimPRMD object}
\usage{
\method{print}{summary.SimPRMD}(x, ...)
}
\arguments{
\item{x}{summary.SimPRMD object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays a useful description of a summary.SimPRMD object
}
