% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prob}
\alias{prob}
\title{Toxicity probability matrix}
\format{An object of class \code{array} of dimension 4 x 3 x 6 x 6 x 3 x 5.}
\usage{
prob
}
\description{
A 6 dimension array providing the toxicity probability of different
scenarios. The dimension is 4 3 6 6 3 5 which represents scenario, cycle
effect, dose level, cycle number, toxicity type, and tox grade. Scenarios are
ordered in a way that the first scenario is MTD = dose 2, the second is MTD =
dose 3, the third is MTD = dose 4, and the fourth is MTD = dose 5. The three
Cycle effect trend are decreasing, flat, and increasing toxicity trend over
cycles. There are 6 doses, 6 cycles, 3 toxicity types and 5 toxicity grades.
}
\keyword{datasets}
