% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SimPRMD.R
\name{summary.SimPRMD}
\alias{summary.SimPRMD}
\title{Summary a SimPRMD object}
\usage{
\method{summary}{SimPRMD}(object, ...)
}
\arguments{
\item{object}{SimPRMD object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\value{
\item{senerio_sum}{Output \code{senerio_sum} of function \code{SimPRMD}}
\item{eff_sum}{Output \code{eff_sum} of function \code{SimPRMD}}
\item{n.trial}{The number of trials in the simulation}
\item{alloc.perc}{The dose allocation percentage for cycle1}
\item{n.stop}{The number of early stop cases}
\item{m.n.pat}{The average number of patient in each trial}
\item{m.dlt.rt}{dlt rate}
\item{c1_dlt.rt}{Cycle 1 dlt rate}
\item{cs_dlt.rt}{Subsequent cycle (cycle > 1) dlt rate}
\item{alloc.perc}{Dose allocation of cycle 1}
\item{sbsq.alloc}{Dose allocation of subsequent cycles (cycle > 1)}
\item{rec.prec}{The percentage of Recommended doses for cycle 1}
\item{effcy.flag}{Argument \code{effcy.flag} of function \code{SimPRMD}}
\item{DLT.drop.flag}{Argument \code{DLT.drop.flag} of function
\code{SimPRMD}}
}
\description{
Summary a SimPRMD object
}
\examples{
## Check ?SimPRMD for example

}
